added `remove` handler

+++ src/cli_commands/config.rs
@@ -76,9 +76,11 @@ pub fn add_config(new_cfg: CliConfig) -> Result<()> {
     Ok(())
 }
 
-pub fn remove_config() -> Result<()> {
+pub fn config_remove() -> Result<()> {
     let cfg_patch = confy::get_configuration_file_path(CONF_FILE, None)?;
-    std::fs::remove_file(cfg_patch)?;
+    if cfg_patch.exists() {
+        std::fs::remove_file(cfg_patch)?;
+    }
     Ok(())
 }
 
@@ -196,7 +198,7 @@ mod tests {
     #[test]
     #[serial]
     fn test_config_file_update() {
-        setup_temp_home_dir_and_execute(|| {
+        setup_temp_home_dir_and_execute("test_config_file_update", || {
             let cli_config_1 = CliConfig {
                 port: "7888".to_string(),
                 ..Default::default()
@@ -218,7 +220,7 @@ mod tests {
 
     #[test]
     fn test_config_file_remove() {
-        setup_temp_home_dir_and_execute(|| {
+        setup_temp_home_dir_and_execute("test_config_file_remove", || {
             add_config(CliConfig {
                 ..Default::default()
             })
@@ -227,17 +229,26 @@ mod tests {
                 .expect("cannot get config file path");
             assert!(cfg_file.exists(), "config file does not exist");
 
-            remove_config().expect("remove_config failed");
+            config_remove().expect("remove_config failed");
             assert!(!cfg_file.exists(), "config must not exist");
         });
     }
 
-    fn setup_temp_home_dir_and_execute<F>(op: F)
+    #[test]
+    fn test_remove_not_existed_config_file() {
+        setup_temp_home_dir_and_execute("test_remove_not_existed_config_file", || {
+            config_remove().expect("remove_config failed");
+        });
+    }
+
+    fn setup_temp_home_dir_and_execute<F>(prefix: &str, op: F)
     where
         F: FnOnce() -> (),
     {
         let env_home_original = std::env::var("HOME").unwrap();
-        let tmp_dir = tempfile::tempdir()
+        let tmp_dir = tempfile::Builder::new()
+            .prefix(prefix)
+            .tempdir()
             .expect("could not create temporary directory")
             .into_path();
         std::env::set_var("HOME", tmp_dir.to_str().unwrap());
